src: Cleanup schemas (#419)
* src: Cleanup schemas

* docs: Fix Schema description

* code: Fix order of class methods
