+++ src/commands/General/User Settings/userconf.js
@@ -25,7 +25,7 @@ module.exports = class extends Command {
 	show(message, [key]) {
 		const path = this.client.gateways.users.getPath(key, { avoidUnconfigurable: true, errors: false, piece: null });
 		if (!path) return message.sendLocale('COMMAND_CONF_GET_NOEXT', [key]);
-		if (!path.piece.path || path.piece.type === 'Folder') {
+		if (path.piece.type === 'Folder') {
 			return message.sendLocale('COMMAND_CONF_USER', [
 				key ? `: ${key.split('.').map(toTitleCase).join('/')}` : '',
 				codeBlock('asciidoc', message.author.settings.list(message, path.piece))
