+++ typings/index.d.ts
@@ -541,27 +541,24 @@ declare module 'klasa' {
 		public toString(): string;
 	}
 
-	abstract class SchemaBase extends Map<string, SchemaPiece | SchemaFolder> {
+	export class Schema extends Map<string, SchemaPiece | SchemaFolder> {
+		public constructor(path?: string);
 		public readonly configurableKeys: Array<string>;
 		public readonly defaults: ObjectLiteral;
+		public readonly path: string;
 		public readonly paths: Map<string, SchemaPiece | SchemaFolder>;
+		public readonly type: 'Folder';
 		public add(key: string, type: string, options?: SchemaPieceOptions): this;
 		public add(key: string, callback: (folder: SchemaFolder) => any): this;
 		public remove(key: string): this;
-		public get<T = SchemaPiece | SchemaFolder>(key: string | Array<string>): T;
+		public get<T = Schema | SchemaPiece | SchemaFolder>(key: string | Array<string>): T;
 		public toJSON(): ObjectLiteral;
-	}
-
-	export class Schema extends SchemaBase {
-		public constructor(path?: string);
-		public readonly path: string;
 	}
 
 	export class SchemaFolder extends Schema {
 		public constructor(parent: Schema | SchemaFolder, key: string);
-		public readonly parent: Schema | SchemaFolder;
 		public readonly key: string;
-		public readonly type: 'Folder';
+		public readonly parent: Schema | SchemaFolder;
 	}
 
 	export class SchemaPiece {
